fixes for missing urls
